Public API Reference

Constants

Functions

FunctionMaps.affinematrix — Method

Return the matrix A in the affine map Ax+b.

FunctionMaps.affinevector — Method

Return the vector b in the affine map Ax+b.

FunctionMaps.canonicalmap — Method

canonicalmap([ctype::CanonicalType, ]map)

Return an associated canonical map, if any, of the given map.

Optionally, a canonical type argument may specify an alternative canonical map. Canonical maps help with converting between equal maps of different types.

FunctionMaps.codomaintype — Method

codomaintype(m[, T])

What is the type of a point in the codomain of the function map m?

The second argument optionally specifies a domain type T.

FunctionMaps.diffvolume — Method

diffvolume(m[, x])

Compute the differential volume (at a point x). If J is the Jacobian matrix, possibly rectangular, then the differential volume is sqrt(det(J'*J)).

If the map is square, then the differential volume is the absolute value of the Jacobian determinant.

FunctionMaps.domaintype — Method

domaintype(m)

What is the expected type of a point in the domain of the function map m?

FunctionMaps.inverse — Method

inverse(m[, x])

Return the inverse of m. The two-argument function evaluates the inverse at the point x.

FunctionMaps.isaffinemap — Method

isaffinemap(m)

Is m an affine map?

If m is affine, then it has the form m(x) = A*x+b.

See also: affinematrix, affinevector.

FunctionMaps.isequalmap — Method

isequalmap(map1, map2)

Are the two given maps equal?

FunctionMaps.islinearmap — Method

islinearmap(m)

Is m a linear map?

FunctionMaps.isrealmap — Method

isrealmap(m)

Is the map real-valued?

A map is real-valued if both its domain and codomain types are real.

FunctionMaps.jacdet — Method

jacdet(m[, x])

Return the determinant of the jacobian as a map. The two-argument version evaluates the jacobian determinant at a point x.

FunctionMaps.jacobian — Method

jacobian(m[, x])

Return the jacobian map. The two-argument version evaluates the jacobian at a point x.

FunctionMaps.leftinverse — Method

leftinverse(m[, x])

Return a left inverse of the given map. This left inverse mli is not unique, but in any case it is such that (mli ∘ m) * x = x for each x in the domain of m.

The two-argument function applies the left inverse to the point x.

FunctionMaps.numtype — Method

The numeric element type of x in a Euclidean space.

FunctionMaps.prectype — Method

prectype(x[, ...])

The floating point precision type associated with the argument(s).

FunctionMaps.rightinverse — Method

rightinverse(m[, x])

Return a right inverse of the given map. This right inverse mri is not unique, but in any case it is such that (m ∘ mri) * y = y for each y in the range of m.

The two-argument function applies the right inverse to the point x.

Types

FunctionMaps.AffineMap — Type

AffineMap{T} <: AbstractAffineMap{T}

The supertype of all affine maps that store `A` and `b`.

Concrete subtypes differ in how A and b are represented.

FunctionMaps.AffineMap — Method

AffineMap(A, b)

Return an affine map with an appropriate concrete type depending on the arguments A and b.

Examples

julia> AffineMap(2, 3)
x -> 2 * x + 3

FunctionMaps.ConstantMap — Type

The supertype of constant maps from T to U.

FunctionMaps.IdentityMap — Type

Supertype of identity maps.

FunctionMaps.LinearMap — Type

LinearMap{T} <: AbstractAffineMap{T}

The supertype of all linear maps y = A*x.

Concrete subtypes may differ in how A is represented.

FunctionMaps.Map — Type

A Map{T} is a map of a single variable of type T.

FunctionMaps.MapRef — Type

MapRef(m)

A reference to a map.

In a function call, MapRef(x) can be used to indicate that x should be treated as a map, e.g., foo(x, MapRef(m)).

FunctionMaps.ProductMap — Type

A product map is diagonal and acts on each of the components of x separately: y = f(x) becomes y_i = f_i(x_i).

FunctionMaps.Translation — Type

A Translation represents the map y = x + b.

FunctionMaps.UnityMap — Type

The unity map f(x) = 1.

FunctionMaps.ZeroMap — Type

The zero map f(x) = 0.